+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ast/ApexCommentBuilder.java
@@ -19,7 +19,6 @@ import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.Token;
 
 import net.sourceforge.pmd.annotation.InternalApi;
-import net.sourceforge.pmd.lang.apex.internal.AntlrVersionCheckSuppression;
 import net.sourceforge.pmd.lang.document.TextDocument;
 import net.sourceforge.pmd.lang.document.TextRegion;
 
@@ -31,8 +30,6 @@ final class ApexCommentBuilder {
     private final CommentInformation commentInfo;
 
     ApexCommentBuilder(TextDocument sourceCode, String suppressMarker) {
-        AntlrVersionCheckSuppression.initApexLexer();
-
         this.sourceCode = sourceCode;
         this.commentInfo = extractInformationFromComments(sourceCode, suppressMarker);
     }